Added new operation codes to the emulator

+++ include/memory.h
@@ -11,8 +11,8 @@
 class Memory {
   public:
     Memory();
-    std::uint8_t getAddress(const std::uint16_t index) const;
-    void setAddress(const std::uint16_t index, const std::uint8_t value);
+    std::uint8_t getAddress(const std::uint32_t index) const;
+    void setAddress(const std::uint32_t index, const std::uint8_t value);
     void loadFont(const Font &font);
     void loadROM(const std::string &name);
     void push(std::uint16_t address);
